Remove `debugging_connector.py` (#10113)
